(nds32) Fix ABI issues as the size of variables is larger than 4.
Gdb intends to align register number to the size of variables.
For example, if the size of variables is 16 and $r0 is occupied, gdb
will use $r4 as next usable register.  However, GCC just align the
register number to even-numbered.  So, callee will expect the variable
be stored in ($r2:$r5) instead of ($r4:$r5 + stack).